chatwidget: support multiline topics and fix pane sizes on re-map.
The topic display is changed to be a small text widget rather than an entry
widget with automatic scrolling to handle channels with long topic messages.
On redisplay of a tab the panes would revert to default sizes due to the <Map>
binding being left in place. Corrected to remove the <Map> binding once used
to set the default pane sizes.